% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_transactions.R
\name{process_transactions}
\alias{process_transactions}
\title{Process transactions}
\usage{
process_transactions(
  transactions,
  seeds = NULL,
  verbose = FALSE,
  tol_amount = 1e-05
)
}
\arguments{
\item{transactions}{data.frame or filename}

\item{seeds}{character, path to seeds folder of the dbt project, default dirname(transactions)}

\item{verbose}{logical, print diagnostic messages}

\item{tol_amount}{numerical, tolerance for position sizes}
}
\value{
NULL, used for its side effects
}
\description{
Takes "raw" transaction data (isin, name, date, size, amount, type, portfolio, broker) and
generates three csv -- cash, active_positions, closed_trades -- in the provided seeds folder.
}
\details{
Is able to handle different portfolios. 

closed_trades: Sell Orders are matched with the oldest Buy Order available. 
Dividends are ignored for now.

cash: is basically the cumulated sum of the transactions. Note that in
trade republic there may be differences between "transactions" in the app and
account statements. It seems account statements are the "real" thing.
}
