% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_universe.R
\name{process_universe}
\alias{process_universe}
\title{Process universe}
\usage{
process_universe(univ_pdf, seeds, verbose = FALSE)
}
\arguments{
\item{univ_pdf}{path to LS trade universum}

\item{seeds}{character, path to seeds folder of the dbt project, default dirname(transactions)}

\item{verbose, }{logical, print diagnostic messages}
}
\value{
NULL, used for its side effects
}
\description{
Takes PDF from Lang & Schwarz, i.e. https://www.ls-x.de/media/lsx/stammdaten20250418.pdf
and converts it to data.frame
}
